requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
python-dotenv==1.0.1
